add org-recur-agenda-mode, add keymaps, fix turning on/off modes
Fixed a few issues having to do with turning on/off the mode in org-agenda. To
do that, I added org-recur-agenda-mode, which was needed anyway -- the
highlighting scheme is different in org-agenda, and users may want to add
different bindings to the keymaps (as in the suggested config in the Readme).
Speaking of which, I added the missing keymaps.